requirements.txt
flask
pymongo
mongoengine
flask-wtf
python-dotenv
flasgger
flask_sqlalchemy
sqlalchemy.orm
typing
pytest
Jinja2
flask-mail
flask-login
flask-bcrypt
bcrypt
itsdangerous
pyopenssl
secrets